+++ apps/api/src/scrapers/interface.ts
